requirements.txt
# Django and REST Framework
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.28.0

# MongoDB integration (using pymongo directly instead of djongo)
pymongo==4.10.1

# Google Gemini AI
google-generativeai==0.8.3

# Machine Learning
scikit-learn==1.6.0
numpy==2.2.1
pandas==2.2.3
joblib==1.4.2

# Environment management
python-decouple==3.8

# Development and testing
pytest==8.3.4
pytest-django==4.9.0
hypothesis==6.122.2

# Security and utilities
cryptography==44.0.0
requests==2.32.3